+++ db/schema.rb
@@ -10,7 +10,8 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2018_07_19_071704) do
+ActiveRecord::Schema.define(version: 2018_09_05_073314) do
+
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -68,6 +69,12 @@ ActiveRecord::Schema.define(version: 2018_07_19_071704) do
     t.index ["user_id"], name: "index_orders_on_user_id"
   end
 
+  create_table "providers", force: :cascade do |t|
+    t.text "name", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
   create_table "service_categories", force: :cascade do |t|
     t.bigint "service_id"
     t.bigint "category_id"
@@ -76,6 +83,15 @@ ActiveRecord::Schema.define(version: 2018_07_19_071704) do
     t.datetime "updated_at", null: false
     t.index ["category_id"], name: "index_service_categories_on_category_id"
     t.index ["service_id"], name: "index_service_categories_on_service_id"
+  end
+
+  create_table "service_opinions", force: :cascade do |t|
+    t.integer "rating", null: false
+    t.text "opinion"
+    t.bigint "order_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["order_id"], name: "index_service_opinions_on_order_id"
   end
 
   create_table "services", force: :cascade do |t|
@@ -89,8 +105,11 @@ ActiveRecord::Schema.define(version: 2018_07_19_071704) do
     t.decimal "rating", precision: 2, scale: 1, default: "0.0", null: false
     t.text "connected_url"
     t.boolean "open_access", default: false
+    t.bigint "provider_id"
+    t.integer "service_opinion_count", default: 0
     t.index ["description"], name: "index_services_on_description"
     t.index ["owner_id"], name: "index_services_on_owner_id"
+    t.index ["provider_id"], name: "index_services_on_provider_id"
     t.index ["title"], name: "index_services_on_title"
   end
 
@@ -113,5 +132,6 @@ ActiveRecord::Schema.define(version: 2018_07_19_071704) do
   end
 
   add_foreign_key "order_changes", "users", column: "author_id"
+  add_foreign_key "services", "providers"
   add_foreign_key "services", "users", column: "owner_id"
 end